Ensure that both ways to get repos by ids are tested
In 222a970 I added a way to get repos by ids using format ids=1,2,3. It
turned out that getting by ids was already supported with
ids[]=1&ids[]=2 format, but it wasn't tested. I added tests for both
methods and left the comma separated ids list as we already support it
in other places.