Overwrite database pool size for logs database
Logs database config is set in travis keychain and is set to a
concurrency much higher than what we set in heroku config. I don't want
to change it globally, because it may need to be different for other
apps, so I'm changing it here just before we connect to the db.