Fix failing specs when running Ruby 2.3.0
There is a subtle difference in behaviour of splat operator between
different Ruby versions. This was making the specs fail on Ruby 2.3.0,
while working fine under older Ruby versions such as 2.1.5.
See this gist for a minimal working example:
https://gist.github.com/piotrdz/f63aa302cb0bfbf7e54b5c4414d3f6a1

This fix changes the calling convention of Travis::API::V3::Result
constructor to work correctly for both versions of Ruby.