Autoloading from lib
It's nicer to have service classes in `app/**`, but Rails' autoloading
ties the class names too closely to the directory structure for this
to be useful :(

This commit add `lib` to the autoload path, which allows us to
structure our service classes as we want.

Also note the modules – if the `Services` module has never been
created before, we can't just do `class Services::Foo::Bar` –
Ruby doesn't know what that means.

For more reading on the problems with Rails and autoloading, check
this post out:
http://urbanautomaton.com/blog/2013/08/27/rails-autoloading-hell/