Allow Range and Content-Length headers
The `Content-Length` header is needed so clients can potentially determine whether they should avoid downloading the
entire log response when it its size would negatively impact performance/UX.

The `Range` header is needed for similar reasons, as I would like to investigate if we can start fetching ranges of bytes
from S3 instead of the entire log at once. This would allow the client to implement a type of infinite scrolling that could intelligently
request the relevant portions of the log based on the overall `Content-Length`.