Use uniform param names in settings endpoints
It came to my attention when writing the docs for these endpoints
that trying to accomodate the previous `setting.something` params
is not worth the hassle now that the resource is named `user_setting`.

This switches to `user_setting.something` but still allows
`setting.something` so that we can deploy this without breaking
travis-web.