Ensure repo id is always set as additional attribute
We found an error where updating a key pair via PATCH resulted
in a rendering error. This was traced to the repo id on the key
pair being nil, because the create service was not setting it.

This also adds tests to cover that case, for key pairs and
env vars.