Run customerio in background
At the moment we need to wait for a customerio's response in order to
log a user in. I think that this might be causing problems for some
users. I'm not 100% certain that this is customerio's problem, but I
think that we should be running it in the background anyway. It's not
crucial for signing in, so we shouldn't be making a handshake call
longer than it needs to be.

This commit creates a customerio worker, which will run the needed
request using sidekiq.