Log GitHub's response when token is missing
For some users we get a response that has a 200 status, but there's no
token. This commit adds logging of the response's body and headers in
order to get more information about the situation. This should be safe
as it will only output the body when token can't be found.