Create User#tokens record on sign up
We still use user tokens that are stored in `tokens` table to
authenticate status images and cc.xml requests. Because of that
travis-web does a check for a token when it fetches the user. Recently
we started creating new user records outside of auth flow - we're basing
it on GitHub payloads. It means that some of the users may not have a
token available. We don't need tokens for users that never signed in, so
the best thing to do is to create a token when a user tries to sigin in.