Subscribe to private user channel
Since the recent switch to per-user channels we always use public user
channel on .org and private user channel on .com. This doesn't make a
lot of sense, especially that when we merge the platforms, we will stay
with only private user channel. To unify the behaviour and simplify the
code in other places, this commit changes API to subscribe user to both
public and private user channels, which will allow to seamlessly switch
the rest of the apps to use private user channel (and then we can drop
public user channel).