Sort branches.last_build by last_build.id (#560)
After I committed 755b831 I realised that this is still not right. When
you restart a job and cancel it right away, started_at will be set to
null, so even though the build might be new, the branch will be put at
the end of the results. Even if we defaulted to the previous behaviour
of NULLs going to the beginning, it would still not be entirely right,
because an old branch with a cancelled build could come up on top.
Sorting by id seems like a best way to achieve what we want, ie. sort by
the most recent build.