Add eql_json rspec matcher
In API V3 specs we often compare big JSON objects and when 2 objects
don't match, we get a message that's not very helpful: expected (...big
pile of JSON here...) to be equal to (...another big pile of JSON
here...).

This commit introduces eql_json matcher that will return a better
message. For example given the following expectation:

    actual = {
      'job' => {
        'id' => 1,
        'number': '10.1',
        'repository_id': 1
      }
    }

    expect(actual).to eql_json({
      'job' => {
        'id' => 1,
        'number': '11.1',
        'state': 'passed'
      }
    })

The message will be:

    Failure/Error: expect(actual).to eql_json({
      expected to match JSON:
        missing job.state == passed
        entries @job.number do not match, expected: 11.1, actual: 10.1
        extra entry job.repository_id == 1

This makes it much easier to fix any potential problems.