Return updated_at field for jobs and builds
API responses and live updates (delivered by Pusher) may sometimes come
to the client out of order. That's why we want to have a way to version
records to be able to compare them and apply only the newest updates.
updated_at column seems like a good way to check record's version, so
this commit adds it to job's and build's payload.

We might consider adding it to other records as well, but for now these
are two most important tables.