Fix fetching accounts when names are duplicated
find_user_accounts service that we use to fetch accounts and
repositories names was querying accounts by getting a list of user
repositories and then using an owner_name attribute to fetch
organizations. This method will work only if we don't have any
duplicates in the database. Duplicates can happen in multiple scenarios,
for example when an organization is deleted from GitHub and a new one is
created. We don't delete old organizations at the moment, so fetching by
name will get both records.

This commit changes the code to fetch organizations by ids rather than
by names.

In theory we could use memberships to fetch organizations (so basically
do `SELECT organization_id FROM memberships WHERE user_id = ?`), but I'm
not sure if this change would have any side effects, so I would have to
check how we store memberships and how exactly memberships work on GitHub.
So for now I prefer to leave the way this code works unchanged, and only
fix the main issue with it.