Fetch the newest organization and user (#628)
We sometimes might have more than one User or Organization record with
the same login in the DB. At the moment we don't invalidate them, which
means that there's a possibility that a user will get repositories from
an older organization. This doesn't really pose any security threat,
because we filter the repositories with a data from permissions table,
but it can result in confusing bugs for the users.

This commit changes queries used to fetch user and organization records
by logins to fetch records with the newest ids. This might not be 100%
accurate, so in the long run we should start invalidating old records,
but it should work in majority of cases (if not all cases).