Fix fetching accounts for a user
When fixing a bug in fetching accounts I actually introduced another
problem that I caught only when testing on staging. The previous code
was searching for Organizations also using a user id, which could result
in fetching an Organization with the same id as a user, even thought the
organization might not be related to the user at all. This commit fixes
the problem.