Fix filtering by slug with uppercase letters
When we were searching by slug we were querying the database for an
owner name and a repo name with all characters converted to lower case.
But the filter string was not converted, which means that it needed to
be in all lower case letters already. This commit fixes the problem by
converting a filter string to lower case as well.