Introduce V2.1 API (#639)
The purpose of V2.1 API is to change APIs behaviour in context of
authentication error codes for responses that may return public data
without breaking v2. When using V2 on travis-ci.com, we will return 401
status for most of the endpoints. In preparation for enabling public
repository access on travis-ci.com we realised that this will not work
well with endpoints that should be publicly accessible. For example when
fetching a build, we should either return a build with a 200 response or
return 404 when a build doesn't exist or when a user doesn't have access
to the build. On API V2 on travis-ci.com, however, we will return a 401
response if a user is not authenticated. Without API V2.1 we would have
to break that behaviour and we're not sure if anyone's using it.

So in essence API V2.1 changes the status code we return for requests
that previously needed a valid auth token on API V2.