[WIP] Add enterprise license endpoint (#573)
* Add EnterpriseLicense endpoint

For now this returns just some sample data

* Render actual number of seats retrieved from Replicated

* Fix Replicated API calls to use actual path and hostname

The Replicated Integrations API hostname gets resolved via the
REPLICATED_INTEGRATIONAPI environment variable. This gets automatically
defined within the Enterprise container and resolves to an IP address
with protocol and port, e.g. https://10.169.183.13:9880

* Returns active_user count for Enterprise license

* Fix broken license file

* Generate test data and fix broken query

* Define enterprise_license route in test env and if enterprise is defined in travis-config

* Add active_users in available attributes constant

* Introduce enterprise method for routes dsl

This method is used to conditionally define endpoints which are only
available in Travis CI Enterprise. It defines the endpoint if we're in
testing mode or if travis-config has the enterprise flag being set

* We also return the expiration_time from replicated's response

* Limit enterprise user query for the last year

This is dependent on the expiration date we get from Replicated.

* update service, query, spec

* update metriks gem to our forked version based on hdrhistogram

* update to branch of travis-support that does not include travis-metrics

so that we can reference the travis-metrics from its own gem

* update travis-support dep

* normalize librato config

* fix code accessing Travis.config.metrics

* update to master branch of travis-support

* Introduce V2.1 API (#639)

The purpose of V2.1 API is to change APIs behaviour in context of
authentication error codes for responses that may return public data
without breaking v2. When using V2 on travis-ci.com, we will return 401
status for most of the endpoints. In preparation for enabling public
repository access on travis-ci.com we realised that this will not work
well with endpoints that should be publicly accessible. For example when
fetching a build, we should either return a build with a 200 response or
return 404 when a build doesn't exist or when a user doesn't have access
to the build. On API V2 on travis-ci.com, however, we will return a 401
response if a user is not authenticated. Without API V2.1 we would have
to break that behaviour and we're not sure if anyone's using it.

So in essence API V2.1 changes the status code we return for requests
that previously needed a valid auth token on API V2.

* Revert "Introduce V2.1 API (#639)"

This reverts commit 7e3cda68b426502b3b8a9cd3af3913de8b115b1b.

We want to first try it from a branch and only then merge to master

* remove debug output

* fix indentation

* use redis-server

* remove comments

* add check for replicated ENV VAR

* add ssl to faraday request

* remove comment

* update active users query