Fix for the sec-incident-20180214 (#4)
log.token was supposed to allow to authenticate only log requests.
Because it used existing log authentication logic, which in turn used
job visibility check, it also allowed to view jobs. This commit fixes
the problem by doing an explicit check for log visibility instead of
relying on the job_visible? check.