Fix some V2 endpoints to not include public stuff on .com (#655)
After switching from travis-pro-api to travis-api to serve traffic on
Travis CI Pro it turned out that there're differences in how certain
endpoints return the data. In travis-pro-api we were scoping all of the
endpoints to return only the stuff that belongs to a user. In travis-api
we return records to which a user has access to *and* public data. This
makes sense in most of the places, but not for `/repos`, `/builds` or
`/jobs`. The assumption there is that you get only your data when you do
an authenticated request, not your data and some random public records.