Allow to hide routes from the index
Sometimes it may be necessary to hide routes from the index, for example
when working on an experimental features. This commit adds a `hide`
method that let's you hide a route.

I went with a kind of a functional approach of `hide(get(:foo))`,
because I didn't want to change too many signatures (ie. adding options
to routes would be more complicated).