Remove eagerloading of :sender
This was causing AR errors when using sort_by=default_branch.last_build
as a filter on /v3/owner/[owner]/repos calls. I'm not convinced this is
a perfect solution, but after an hour of being on org-production I saw
no discernible change in DB load or performance, and it allows the
queries in question to resolve.

Error in sentry:
https://sentry.io/travis-ci/org-api-production/issues/571746986/?query=request_id:d4e62c95-b674-4f17-b967-1df4e70757b7

Issue:
https://github.com/travis-pro/team-teal/issues/2709