Hide migrating repositories behind a feature flag
We don't want to make migrating repositories available to everyone just
yet, so we're hiding it behind a feature flag.

This commit also slightly changes how we check for admin access. Before
the change the migrate service was checking for a migrate permission,
which only required write access and then it was searching for an admin,
returning 404 when no admin could be found. This might be confusing to
the users and I think it's better to explicitly check for admin access
for the migrate permission.