Make sure that branches use upsert
It seems that on production `connection.index_exists?` for a unique
partial index returns false (it might be due to different postgresql
version). It also turned out that using upsert in older versions of
postgresql need to use an "index_predicate", which means adding `WHERE
unique_name` to the INSERT statement in order to actually use the unique
index.

In this commit I remove the index check - index is already deployed to
all platforms - and I also fix the query to use predicate.