Fix Repository#default_branch method
When changing the method to use upsert I missed some things about how it
works, mainly that it checks if a branch returned from
Repository#create_branch is a new record and if it is, a last_build will
be attached there. When using upsert a record will be always fetched
from the database, so I needed to change how the method works.