Fix another bug that creates duplicated branches
The fix for creating branch duplicates that I shipped to API was
wrong. It only fixed situation for new branches, but was still creating
duplicates for existing branches. The problem is that upsert uses
`unique_name` column and not `name` column. The former is not populated
for older records, so upsert won't work then. Unfortunately due to a way
I wrote a trigger test suite was passing without any problems.