Fix Repository#find_or_create
The upsert statement needs an index, so it's best to ensure that the
unique index is already created in the database, that way the code will
work fine even if the column is added, but the index hasn't been created
yet.

Another problem is that because we set unique_name only to new or
updated rows, so there is potentially a chance to create a duplicate of
an existing row where the existing row would be left with unique_name
set to NULL and a new row's unique_name would be set to the same value
as name. If that's the case we need to always fetch the row with
unique_name that is not NULL.