Remove managed_by_installation_at filter from webhooks migration script
The 'managed_by_installation_at' filter in the migration was used to
limit the number of repositories to check when running on hosted. This
was necessary due to the large volume of repos to check.

This column is not available in Enterprise 2.2 due to the fact that we
are running a much older migration version. Additionally, the number of
repos to check is much lower in an individual Enterprise instance. For
these reasons we can remove the filter at the cost of additional
runtime.