Rename config values to ease production rollout
Previously, we were renaming the Travis.config logs values referenced,
preferring explicit com/org settings. This would have required us to
update .org as well as .com with the new values.

In case anything were to go wrong and we needed to roll back, that would
have required us to update the config names as well, adding pressure to
the situation and increasing the odds that we make a mistake.

Instead, we can simply add a fallback url which is only used by com. If
we need to roll back, we can do so without issue, as the default
settings from before were not touched.