Redirect badges to .com (#914)
At the moment most of the READMEs of open source projects and some
services (like cargo, npm etc) point to travis-ci.org as a source of
badges. In order to make it easier to transition this commit changes the
behaviour of badges to redirect some of the badges requests to .com. If
a repo is migrated, a request for a badge will be redirected to .com. If
a repo is not found, a request would also be redirected to .com, which
will ensure that new repos on .com will also be available when using the
.org url.