Change default builds sort to created_at:desc,id:desc
We previously could rely on the id being in the order that builds were
created. However, as part of build history migration, we discovered that
this no longer holds.

We import builds via two discrete paths (recent/older) and split those
across queues to ensure that we don't block the import of recent history
of smaller repositories with the potentially massive legacy history of a
larger repository.

The side effect, however, is that as builds are copied over, we no
longer control the id ordering, that is, a build with a lower id does
*not* mean that it was created before a build with a higher id. In the
UI, this leads to us not showing the most recent builds as expected.

Instead of simply adding a new `sort_by` parameter, it makes sense that we
preserve the default expected behavior, but adapted to our new reality.
The id:desc is for the (unlikely) situation where we have identical
`created_at` timestamps, that we should take the "older" record with the
higher id.