Redirect badges also to .org
Some platforms (like crates.io) rely on using only one badges URL as
they can't know if a badge is supposed to be on .org or .com. Some
customers might also want to be able to point their scripts to only one
URL.

This commit adds a badge redirect from .com to .org if a repo is not
migrated, if it's public and if it's not active on .com.