Remove unused method call API for `last_build`
This was a strange case, as I could not find it used anywhere in the app itself. Instead, we now call the method without arguments as it is clear we only want the most recent build, regardless of branch. If this is incorrect, I am not sure how we were enforcing it before, as the `has_one` association did not mention the branch either.