Fix several remaining exceptions related to settings
I honestly don't understand why any of this would be necessary.
AR should, according to error messages I get when attempting to
`serialize :settings`, transparently convert to/from AR models <-> JSON,
but every time I inspect the settings, we are getting the string instead
of a hash.

This work is literally just fixing error messages in tests, but I
haven't been able to get to the bottom of what changed in AR between
versions. Maybe I'm just missing something and there's going to be a
cleaner way to fix this...